+++ packages/changelog-github/src/index.ts
@@ -78,9 +78,10 @@ const changelogFunctions: ChangelogFunctions = {
           pull: prFromSummary,
         });
         if (commitFromSummary) {
+          const shortCommitId = commitFromSummary.slice(0, 7);
           links = {
             ...links,
-            commit: `[\`${commitFromSummary}\`](https://github.com/${options.repo}/commit/${commitFromSummary})`,
+            commit: `[\`${shortCommitId}\`](https://github.com/${options.repo}/commit/${commitFromSummary})`,
           };
         }
         return links;